// packages/functions/src/types/enums.ts
export enum Phase {
  SHORTLIST = 'SHORTLIST',
  NOMINATION = 'NOMINATION',
  WINNER = 'WINNER'
}

export enum UserRole {
  ADMIN = 'ADMIN',
  TESTER = 'TESTER',
  USER = 'USER'
}

export enum AwardsBody {
  ACADEMY_AWARDS = 'ACADEMY_AWARDS',
  GOLDEN_GLOBES = 'GOLDEN_GLOBES',
  CRITICS_CHOICE = 'CRITICS_CHOICE',
  BAFTA = 'BAFTA',
  HCA = 'HCA',
  PGA = 'PGA',
  SAG = 'SAG',
  DGA = 'DGA',
  WGA = 'WGA',
  ADG = 'ADG',
  MAKEUP_GUILD = 'MAKEUP_GUILD',
  CDG = 'CDG',
  ASC = 'ASC',
  MPSE = 'MPSE'
}

export enum EventStatus {
  NOMS_STAGING = 'NOMS_STAGING', // before nominations go public
  NOMS_LIVE = 'NOMS_LIVE', // when everyone can predict nominations
  WINS_STAGING = 'WINS_STAGING', // when nominations are being announced + prepared
  WINS_LIVE = 'WINS_LIVE', // when everyone can predict wins
  ARCHIVED = 'ARCHIVED' // when winners have been announced and such
}

export enum CategoryType {
  FILM = 'FILM',
  PERFORMANCE = 'PERFORMANCE',
  SONG = 'SONG'
}

export enum CategoryName {
  PICTURE = 'PICTURE',
  DIRECTOR = 'DIRECTOR',
  ACTOR = 'ACTOR',
  ACTRESS = 'ACTRESS',
  SUPPORTING_ACTOR = 'SUPPORTING_ACTOR',
  SUPPORTING_ACTRESS = 'SUPPORTING_ACTRESS',
  ORIGINAL_SCREENPLAY = 'ORIGINAL_SCREENPLAY',
  ADAPTED_SCREENPLAY = 'ADAPTED_SCREENPLAY',
  SCREENPLAY = 'SCREENPLAY',
  INTERNATIONAL = 'INTERNATIONAL',
  ANIMATED = 'ANIMATED',
  DOCUMENTARY = 'DOCUMENTARY',
  EDITING = 'EDITING',
  CINEMATOGRAPHY = 'CINEMATOGRAPHY',
  PRODUCTION_DESIGN = 'PRODUCTION_DESIGN',
  COSTUMES = 'COSTUMES',
  MAKEUP = 'MAKEUP',
  VISUAL_EFFECTS = 'VISUAL_EFFECTS',
  SOUND = 'SOUND',
  SCORE = 'SCORE',
  SONG = 'SONG',
  SHORT_ANIMATED = 'SHORT_ANIMATED',
  SHORT_DOCUMENTARY = 'SHORT_DOCUMENTARY',
  SHORT_LIVE_ACTION = 'SHORT_LIVE_ACTION',
  ENSEMBLE = 'ENSEMBLE',
  COMEDY_PICTURE = 'COMEDY_PICTURE',
  COMEDY_ACTOR = 'COMEDY_ACTOR',
  COMEDY_ACTRESS = 'COMEDY_ACTRESS',
  ACTION_PICTURE = 'ACTION_PICTURE',
  YOUNG_ACTOR = 'YOUNG_ACTOR',
  RISING_STAR = 'RISING_STAR',
  DEBUT = 'DEBUT',
  FIRST_SCREENPLAY = 'FIRST_SCREENPLAY',
  BRITISH_PICTURE = 'BRITISH_PICTURE',
  ANIMATED_PERFORMANCE = 'ANIMATED_PERFORMANCE',
  BLOCKBUSTER = 'BLOCKBUSTER',
  ACTING_ACHIEVEMENT = 'ACTING_ACHIEVEMENT',
  FEMALE_DIRECTOR = 'FEMALE_DIRECTOR',
  MALE_DIRECTOR = 'MALE_DIRECTOR',
  INDIE_PICTURE = 'INDIE_PICTURE',
  BREAKTHROUGH = 'BREAKTHROUGH'
}